# api-snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UsageBucketOut": {
        "properties": {
          "cache_read_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Read Tokens"
          },
          "cache_write_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Write Tokens"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "fact_count": {
            "title": "Fact Count",
            "type": "integer"
          },
          "incomplete_fact_count": {
            "title": "Incomplete Fact Count",
            "type": "integer"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "missing_fields": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Missing Fields",
            "type": "object"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "reasoning_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Tokens"
          },
          "reported_cost_usd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Cost Usd"
          },
          "request_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Count"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "unclassified_fact_count": {
            "default": 0,
            "title": "Unclassified Fact Count",
            "type": "integer"
          },
          "unpriced_fact_count": {
            "title": "Unpriced Fact Count",
            "type": "integer"
          },
          "unpriced_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Unpriced Reasons",
            "type": "object"
          },
          "usd_equivalent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd Equivalent"
          },
          "valuation_bases": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Valuation Bases",
            "type": "object"
          }
        },
        "required": [
          "fact_count",
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_write_tokens",
          "reasoning_tokens",
          "total_tokens",
          "request_count",
          "usd_equivalent",
          "reported_cost_usd",
          "missing_fields",
          "unpriced_fact_count",
          "incomplete_fact_count",
          "date"
        ],
        "title": "UsageBucketOut",
        "type": "object"
      },
...
      "UsageEraseIn": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "default": false,
            "title": "Confirmed",
            "type": "boolean"
          }
        },
        "title": "UsageEraseIn",
        "type": "object"
      },
      "UsageEraseOut": {
        "properties": {
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision"
        ],
        "title": "UsageEraseOut",
        "type": "object"
      },
      "UsageGroupOut": {
        "properties": {
          "cache_read_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Read Tokens"
          },
          "cache_write_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Write Tokens"
          },
          "deleted": {
            "default": false,
            "title": "Deleted",
            "type": "boolean"
          },
          "fact_count": {
            "title": "Fact Count",
            "type": "integer"
          },
          "incomplete_fact_count": {
            "title": "Incomplete Fact Count",
            "type": "integer"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "missing_fields": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Missing Fields",
            "type": "object"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "reasoning_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Tokens"
          },
          "reported_cost_usd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Cost Usd"
          },
          "request_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Count"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "unclassified_fact_count": {
            "default": 0,
            "title": "Unclassified Fact Count",
            "type": "integer"
          },
          "unpriced_fact_count": {
            "title": "Unpriced Fact Count",
            "type": "integer"
          },
          "unpriced_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Unpriced Reasons",
            "type": "object"
          },
          "usd_equivalent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd Equivalent"
          },
          "valuation_bases": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Valuation Bases",
            "type": "object"
          }
        },
        "required": [
          "fact_count",
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_write_tokens",
          "reasoning_tokens",
          "total_tokens",
          "request_count",
          "usd_equivalent",
          "reported_cost_usd",
          "missing_fields",
          "unpriced_fact_count",
          "incomplete_fact_count",
          "key"
        ],
        "title": "UsageGroupOut",
        "type": "object"
      },
      "UsageMetricsOut": {
        "properties": {
          "cache_read_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Read Tokens"
          },
          "cache_write_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Write Tokens"
          },
          "fact_count": {
            "title": "Fact Count",
            "type": "integer"
          },
          "incomplete_fact_count": {
            "title": "Incomplete Fact Count",
            "type": "integer"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "missing_fields": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Missing Fields",
            "type": "object"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "reasoning_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Tokens"
          },
          "reported_cost_usd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Cost Usd"
          },
          "request_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Count"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "unclassified_fact_count": {
            "default": 0,
            "title": "Unclassified Fact Count",
            "type": "integer"
          },
          "unpriced_fact_count": {
            "title": "Unpriced Fact Count",
            "type": "integer"
          },
          "unpriced_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Unpriced Reasons",
            "type": "object"
          },
          "usd_equivalent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usd Equivalent"
          },
          "valuation_bases": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Valuation Bases",
            "type": "object"
          }
        },
        "required": [
          "fact_count",
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_write_tokens",
          "reasoning_tokens",
          "total_tokens",
          "request_count",
          "usd_equivalent",
          "reported_cost_usd",
          "missing_fields",
          "unpriced_fact_count",
          "incomplete_fact_count"
        ],
        "title": "UsageMetricsOut",
        "type": "object"
      },
      "UsageOverviewOut": {
        "properties": {
          "as_of": {
            "title": "As Of",
            "type": "integer"
          },
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/UsageGroupOut"
            },
            "title": "Breakdown",
            "type": "array"
          },
          "breakdown_total": {
            "title": "Breakdown Total",
            "type": "integer"
          },
          "catalog": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Catalog",
            "type": "object"
          },
          "history_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "History Status",
            "type": "object"
          },
          "last_observed_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Observed At"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "source_breakdown": {
            "additionalProperties": {
              "$ref": "#/components/schemas/UsageMetricsOut"
            },
            "title": "Source Breakdown",
            "type": "object"
          },
          "summary": {
            "$ref": "#/components/schemas/UsageMetricsOut"
          },
          "sync_state": {
            "$ref": "#/components/schemas/UsageSyncStateOut"
          },
          "timeseries": {
            "items": {
              "$ref": "#/components/schemas/UsageBucketOut"
            },
            "title": "Timeseries",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "unallocated": {
            "$ref": "#/components/schemas/UsageMetricsOut"
          },
          "undated": {
            "$ref": "#/components/schemas/UsageMetricsOut"
          }
        },
        "required": [
          "revision",
          "as_of",
          "timezone",
          "summary",
          "timeseries",
          "breakdown",
          "breakdown_total",
          "undated",
          "unallocated"
        ],
        "title": "UsageOverviewOut",
        "type": "object"
      },
      "UsagePriceIn": {
        "additionalProperties": false,
        "properties": {
          "effective_from": {
            "minimum": 0.0,
            "title": "Effective From",
            "type": "integer"
          },
          "effective_to": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "model": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "price_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Price Id",
            "type": "string"
          },
          "provider": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "rates": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Rates",
            "type": "object"
          }
        },
        "required": [
          "price_id",
          "provider",
          "model",
          "effective_from",
          "rates"
        ],
        "title": "UsagePriceIn",
        "type": "object"
      },
      "UsagePriceOut": {
        "properties": {
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision"
        ],
        "title": "UsagePriceOut",
        "type": "object"
      },
...
      "UsageSyncStateOut": {
        "properties": {
          "discard_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Discard Reasons",
            "type": "object"
          },
          "discarded_event_count": {
            "default": 0,
            "title": "Discarded Event Count",
            "type": "integer"
          },
          "gap_count": {
            "default": 0,
            "title": "Gap Count",
            "type": "integer"
          },
          "scope": {
            "default": "daemon",
            "title": "Scope",
            "type": "string"
          },
          "source_count": {
            "default": 0,
            "title": "Source Count",
            "type": "integer"
          },
          "status": {
            "default": "unavailable",
            "title": "Status",
            "type": "string"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Status Counts",
            "type": "object"
          }
        },
        "title": "UsageSyncStateOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Mandri Daemon",
    "version": "0.1.0b0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/usage/overview": {
      "get": {
        "operationId": "usage_overview",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "root_session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Root Session Id"
            }
          },
          {
            "in": "query",
            "name": "project_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Path"
            }
          },
          {
            "in": "query",
            "name": "harness",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Harness"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "billing_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Billing Mode"
            }
          },
          {
            "in": "query",
            "name": "from_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Ms"
            }
          },
          {
            "in": "query",
            "name": "to_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Ms"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "title": "Timezone",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Deleted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_descendants",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Descendants",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "default": "model",
              "enum": [
                "session",
                "project",
                "model",
                "harness",
                "billing_mode"
              ],
              "title": "Group By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/prices": {
      "post": {
        "operationId": "add_usage_price",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsagePriceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsagePriceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Price",
        "tags": [
          "usage"
        ]
      }
    },
...
    "/v1/usage/sessions/{session_id}/erase": {
      "post": {
        "operationId": "erase_session_usage",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageEraseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageEraseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Erase",
        "tags": [
          "usage"
        ]
      }
    }
  }
}